Isolate manifest expansion tracing tests
Remove the raw entry name from manifest condition trace events and make
when conditions evaluate against entry vars before overlaying foreach item
and index locals.

Split the large manifest expansion test file into smaller modules and use
thread-local tracing captures so filtering debug assertions do not share
state across parallel tests.
